requirements.txt
streamlit>=1.30
requests>=2.31
pdfplumber>=0.10
python-docx>=1.1
python-dotenv>=1.0
